TAP5-2749: use existing test aplication instead of creating new one
