+++ tapestry-jpa/src/test/java/org/example/app6/pages/Login.java
@@ -1,0 +1,57 @@
+// Copyright 2023 The Apache Software Foundation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.example.app6.pages;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+import org.example.app6.AppConstants;
+import org.example.app6.entities.User;
+import org.apache.tapestry5.annotations.InjectPage;
+import org.apache.tapestry5.annotations.Property;
+import org.apache.tapestry5.ioc.annotations.Inject;
+import org.apache.tapestry5.services.ApplicationStateManager;
+
+public class Login
+{
+	@Property
+	private String firstname;
+	
+	@Property
+	private String lastname;
+	
+	@Inject
+	@PersistenceContext(unitName = AppConstants.TEST_PERSISTENCE_UNIT)
+	private EntityManager entityManager;
+	
+	@Inject
+	private ApplicationStateManager applicationStateManager;
+	
+	@InjectPage
+	private Sign index;
+	
+	Object onSuccessFromLogin()
+	{
+		entityManager.getTransaction().begin();
+		User user = new User();
+		user.setFirstName(firstname);
+		user.setLastName(lastname);
+		entityManager.persist(user);
+		entityManager.getTransaction().commit();
+		
+		applicationStateManager.set(User.class, user);
+		return index;
+	}
+}
