--- tapestry-jpa/src/test/java/org/example/app7/entities/User.java
@@ -1,48 +1,0 @@
-// Copyright 2023 The Apache Software Foundation
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.example.app7.entities;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-
-@Entity
-public class User
-{
-	@Id
-	private String username;
-	
-	private String password;
-	
-	public String getPassword()
-	{
-		return password;
-	}
-	
-	public void setPassword(String password)
-	{
-		this.password = password;
-	}
-	
-	public String getUsername()
-	{
-		return username;
-	}
-	
-	public void setUsername(String username)
-	{
-		this.username = username;
-	}
-	
-}
